Refactor naistenhelsinki app models and regenerate migrations.
Refactor the model classes of naistenhelsinki app. Create page model for the
page that has all the places located on a map. Use OpenStreetMaps for Place
map admin widget and set default coordinates to Helsinki instead of the center
of Atlantic Ocean (0,0 coordinates).

Regenerate all naistenhelsinki migrations because the model renaming.