Improve first_p filter
Previously the `first_p` filter got the data from the beginning of the value string to the next closing </p> tag, but this causes issue as the value can also contains unwanted tags before the first p-tag so that needs to take consider as it can affects to page rendering.

The `first_p` currently retrieves first p-tag with its content by using regex and it removes all the attributes from the p-tag.